[aps,tests] add casual options for transformer
